% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_INLAB_data.R
\name{read_INLAB_data}
\alias{read_INLAB_data}
\title{import and tidy data from INLAB generated excel files}
\usage{
read_INLAB_data(file.pattern, dir.path = NULL)
}
\arguments{
\item{file.pattern}{A string pattern to match the files to be read}

\item{dir.path}{A string path to the directory containing the files to be read}
}
\value{
A data.table containing the imported data, together with calculated age an single date values
}
\description{
import and tidy data from INLAB generated excel files
}
\examples{
read_INLAB_data("BlutDxI", dir.path = "/home/olli/R_local/labStat/")

}
