% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZLM_RI.R
\name{ZLM_RI}
\alias{ZLM_RI}
\title{takes an analyte vector from read_INLAB_data function and calculates the reference interval using the refineR package.}
\usage{
ZLM_RI(data, analyte, NBootstrap = 0, seed = 8173)
}
\arguments{
\item{data}{A data.table containing the imported data, together with calculated age an single date values}

\item{analyte}{A string with the name of the analyte to calculate the reference interval for, must be contained within the data.table from read_INLAB_data}

\item{NBootstrap}{Number of bootstrap samples to use for the calculation}

\item{seed}{Seed for the random number generator}
}
\value{
list of reference interval values
}
\description{
takes an analyte vector from read_INLAB_data function and calculates the reference interval using the refineR package.
}
\examples{
ZLM_RI(data = read_INLAB_data("BlutDxI", dir.path = "/home/olli/R_local/labStat/"), analyte = "Kreatinin")
}
